Improves bind_addr
Instead of retrieving the first inet address, it retrieves the
intersection of all inet addresses and those listed as server IPs.
This way, we make sure that the bind_addr is something the other
nodes are expecting.

This came about as Linode binds private IP address to the same
interface. The first inet address was always the public address,
but we wanted the private address.